requirements/base.txt
asgiref==3.2.10
Django==3.1.2
djangorestframework==3.12.1
drf-nested-routers==0.92.1
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.4.1
python-decouple==3.3
rest_framework_swagger